requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.4
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
djangorestframework_simplejwt== 4.7.0
requests_mock>=1.9.0
